requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
shap>=0.44.0
loguru>=0.7.2
httpx>=0.26.0
pyarrow>=15.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0